+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- show option to `listen` command.
+- `show` option to `listen` command.
+- `weight` option to the `blink` command.
+- `priority` option to the `blink` command.
+- Mutually exclusive `txt` and `property` options to the `blink` command.
 
 ## [0.2.0] - 2021-01-22
 ### Changed